extract response data logic
